This commit allows to display emoji characters both in the left plane and in the tittle of the chat area.
I've removed most of the tags in the Emojis::textToEmojiText method in order to maintain the vertical alignment of the tittle in the chat area, along with the <p dir="..."> tag (this one inserted some unwanted vertical space). In my tests, the alignment is fine, but I may have broken something in RTL languages.